Mostly finished semester changing.
First off. The way you get to the webpage is go to the admin site and
click on seelio. Since we decided we weren’t going to do any
customization with that I decided to take it over. Sadly, my drawing
skills are shit and hopefully one of you can make a new logo for that
on the admin site.

Now, into the details of what I did. I found out that copying the
folders was not going to work because they are bound to a port and
dynamically changing that was also going to be a pain in the ass. When
I figured this out I also realized that the data in the folder was
useless to copy over and save. The only thing that will change
dynamically per semester is going to be the images which they can more
than easily store on their own computers. Everything worth storing is
going to be in the database.

So with that I just decided to prepend all data inputted during the
current semester with the semester name and year. so submission001
becomes Spring2014submission001. This proves very easy to do and makes
it incredibly easy to grab the keys for a certain semester (by
specifying keys(“Spring2014*) or something like that) So all I did was
add a couple methods in redisDB.py so that the program can set and get
the current semester as well as add a semester to a list of all
previous semesters and get all the previous semesters.